Show favorite location on the map

+++ lib/ui/view_model/weather_detail_view_model.dart
@@ -3,9 +3,9 @@ import 'dart:ui';
 class WeatherDetailViewModel {
   int? id;
   String? name;
-  double? temp;
-  double? tempMin;
-  double? tempMax;
+  int? temp;
+  int? tempMin;
+  int? tempMax;
   String? weatherIcon;
   String? weatherCondition;
   Color? color;
